fix: set load-on-startup for automatically registered Vaadin servlet
When using Vite, DevModeInitializer blocks dev-server startup until a
VaadinServlet is deployed because it needs to get the servlet path to use.
If the container lazily loads servlets, Vite will not start until the first
HTTP request for the Vaadin servlet is received.
This change sets load-on-startup feature for all Vaadin Servlet registered
by the extension, to ensure that Vite is loaded at the startup of the
Web application.
A warning message is shown for custom Vaadin servlets that do not set
loadOnStartup on WebServlet annotation.

Part of vaadin/flow#14479