test: retry count assertion on session invalidation
Session context tests check the counter value hold by a session bean right
after invalidating the Vaadin session. Since the counter is incremented
in the PreDestroy hook, it may happen that when the test checks for the
counter value, the callback has not yet been invoked.
This change retries the counter check for at most on second, to preven
ttest flakyness.

Fixes #126