fix: index optional dependencies (#172) (#173)
Development (e.g. vaadin-dev-server) and optional (e.g. flow-react) dependencies
must not be present in the Vaadin platform indexed, to prevent NoClassDefFound
errors at runtime, for example when running a production build with dev-server
excluded.
This change adds a build step to index optional dependencies only if they are
present in the project configuration.

Part of #158