fix: try using bean manager to instantiate components
Similar to Vaadin routes, also UI components may require dependency injection.
This change update QuarkusInstantiator to try to use BeanManager to get a
component instance, otherwise it fallbacks to the default instantiator.

To be noted that UI component should have `@Dependant` or a proper Vaadin
scope to work correctly with Vaadin; `@ApplicationScoped` and
`@Singleton` must not be used.

Fixes #192