feat: add embedded Vaadin plugin
Add comprehensive Vaadin build tooling integration that embeds the Vaadin
frontend build process directly into the Quarkus build lifecycle, eliminating
the need for external Maveni/Gradle plugin configuration.

Configuration is available through `vaadin.build.*` properties with
comprehensive options for customizing the build process.

It also defines vaadin-dev as a development dependency so that it
will be automatically added when the application is started in
dev mode.

The result of the changes is that an application using Vaadin Quarkus
extension does not need a specific production profile to generate
a production build.

CAVEATS:
- the vaadin-dev dependency version must be hard-coded in the extension
runtime POM file, so if a different version is required for the
application, an explicit dependency must be added to the build tool
configuration.
- the definition of the automatic vaadin-dev should be done in the
platform extension
- The plugin requires information about the project to correctly build
fronted resources; unfortunately the information are not available
at build time unless the quarkus.bootstrap.workspace-discovery flag
is activate. To overcome to this limitation, the extension provides
a CodeGenProvider that collects the required info when executed by
quarkus generate-code and generate-code-tests goals and writes them
into the build folder to be able to read them again in the build steps.