fix: prevent failure when analyzing classes in default package
For native support, the extension inspects the Jandex index to find all classes
under specified packages.
However, the scan fails with an NPE if the index contains classes from the default
package.
This fix adds a null check to prevent the NPE and also marks the vaadinNativeSupport
build step to only be executed during native build.

Fixes #222