fix: add reflection hints for Dependency class (#242)
In previous Flow versions, Dependency class had a toJson method that created
an elemental JsonObject to be sent to the client as JSON. However, the method
has been removed and replaced by Jackson serialization. This causes wrong JSON
serialization in native executables because reflection is required but metadata
is not stored for the Dependency class.

This change adds the missing reflection hint for the Dependency clas.

Fixes #239